refactor: 제출로직 훅화
